pyproject.toml
[tool.poetry]
name = "daily-bible-crawler"
version = "0.1.0"
description = ""
authors = ["joseph <joseph@buzzni.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.3"
playwright = "^1.42.0"
google-auth-oauthlib = "^1.2.0"
google-auth = "^2.28.1"
google-api-python-client = "^2.120.0"
tenacity = "^8.2.3"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
